Split aggregation transforms to own file
