+++ exporters/otlp/otlpmetric/internal/transform/metrics.go
@@ -64,85 +64,3 @@ func metric(m metricdata.Metrics) (*mpb.Metric, error) {
 	}
 	return out, err
 }
-
-func Gauge[N int64 | float64](g metricdata.Gauge[N]) *mpb.Metric_Gauge {
-	return &mpb.Metric_Gauge{
-		Gauge: &mpb.Gauge{
-			DataPoints: DataPoints(g.DataPoints),
-		},
-	}
-}
-
-func Sum[N int64 | float64](s metricdata.Sum[N]) (*mpb.Metric_Sum, error) {
-	t, err := Temporality(s.Temporality)
-	return &mpb.Metric_Sum{
-		Sum: &mpb.Sum{
-			AggregationTemporality: t,
-			IsMonotonic:            s.IsMonotonic,
-			DataPoints:             DataPoints(s.DataPoints),
-		},
-	}, err
-}
-
-func DataPoints[N int64 | float64](dPts []metricdata.DataPoint[N]) []*mpb.NumberDataPoint {
-	out := make([]*mpb.NumberDataPoint, 0, len(dPts))
-	for _, dPt := range dPts {
-		ndp := &mpb.NumberDataPoint{
-			Attributes:        Iterator(dPt.Attributes.Iter()),
-			StartTimeUnixNano: uint64(dPt.StartTime.UnixNano()),
-			TimeUnixNano:      uint64(dPt.Time.UnixNano()),
-		}
-		switch v := any(dPt.Value).(type) {
-		case int64:
-			ndp.Value = &mpb.NumberDataPoint_AsInt{
-				AsInt: v,
-			}
-		case float64:
-			ndp.Value = &mpb.NumberDataPoint_AsDouble{
-				AsDouble: v,
-			}
-		}
-		out = append(out, ndp)
-	}
-	return out
-}
-
-func Histogram(h metricdata.Histogram) (*mpb.Metric_Histogram, error) {
-	t, err := Temporality(h.Temporality)
-	return &mpb.Metric_Histogram{
-		Histogram: &mpb.Histogram{
-			AggregationTemporality: t,
-			DataPoints:             HistogramDataPoints(h.DataPoints),
-		},
-	}, err
-}
-
-func HistogramDataPoints(dPts []metricdata.HistogramDataPoint) []*mpb.HistogramDataPoint {
-	out := make([]*mpb.HistogramDataPoint, 0, len(dPts))
-	for _, dPt := range dPts {
-		out = append(out, &mpb.HistogramDataPoint{
-			Attributes:        Iterator(dPt.Attributes.Iter()),
-			StartTimeUnixNano: uint64(dPt.StartTime.UnixNano()),
-			TimeUnixNano:      uint64(dPt.Time.UnixNano()),
-			Count:             dPt.Count,
-			Sum:               &dPt.Sum,
-			BucketCounts:      dPt.BucketCounts,
-			ExplicitBounds:    dPt.Bounds,
-			Min:               dPt.Min,
-			Max:               dPt.Max,
-		})
-	}
-	return out
-}
-
-func Temporality(t metricdata.Temporality) (mpb.AggregationTemporality, error) {
-	switch t {
-	case metricdata.DeltaTemporality:
-		return mpb.AggregationTemporality_AGGREGATION_TEMPORALITY_DELTA, nil
-	case metricdata.CumulativeTemporality:
-		return mpb.AggregationTemporality_AGGREGATION_TEMPORALITY_CUMULATIVE, nil
-	default:
-		err := fmt.Errorf("unknown temporality: %s", t)
-		return mpb.AggregationTemporality_AGGREGATION_TEMPORALITY_UNSPECIFIED, err
-	}
-}
